+++ packages/bpk-docs/src/routes/Routes.js
@@ -73,6 +73,7 @@ import AlignmentPage from './../pages/AlignmentPage';
 
 import NativeInputPage from './../pages/NativeTextInputPage';
 import NativeTextPage from './../pages/NativeTextPage';
+import NativeButtonPage from './../pages/NativeButtonPage';
 
 import ResourcesPage from './../pages/ResourcesPage';
 
@@ -135,6 +136,7 @@ const Routes = (
         <IndexRedirect to={ROUTES.NATIVE_TEXT} />
         <Route path={ROUTES.NATIVE_INPUT} component={NativeInputPage} />
         <Route path={ROUTES.NATIVE_TEXT} component={NativeTextPage} />
+        <Route path={ROUTES.NATIVE_BUTTON} component={NativeButtonPage} />
       </Route>
       <Route path={ROUTES.UTILITIES}>
         <IndexRedirect to={ROUTES.ALIGNMENT} />